rename file_info.py -> utils.py
